\name{summary}
\alias{summary}
\alias{summary.Bootstrap.Stage.2.NormScore}
\alias{summary.ExploreData}
\alias{summary.ICC}
\alias{summary.Stage.2.NormScore}
\alias{summary.GLT}
\alias{summary.Stage.1}
\alias{summary.Stage.2.NormTable}
\alias{summary.Sandwich}
\alias{summary.CheckFit}
\alias{summary.Tukey.HSD}

\title{Summary}
\description{summary}
\usage{
\method{summary}{ExploreData}(object, ..., Object)
\method{summary}{ICC}(object, ..., Object)
\method{summary}{Stage.2.NormScore}(object, ..., Object)
\method{summary}{GLT}(object, ..., Object, Assume.Homoscedasticity)
\method{summary}{Stage.1}(object, ..., Object)
\method{summary}{Stage.2.NormTable}(object, ..., Object)
\method{summary}{Sandwich}(object, ..., Object)
\method{summary}{CheckFit}(object, ..., Object)
\method{summary}{Bootstrap.Stage.2.NormScore}(object, ..., Object)
\method{summary}{Tukey.HSD}(object, ..., Object)
}
\arguments{
\item{object}{An object for which a summary is desired.}
\item{...}{Other arguments.}
\item{Object}{An object for which a summary is desired.}
\item{Assume.Homoscedasticity}{Logical. Should homoscedasticity be assumed in the GLT procedure?.}
}
\value{No return value, called for side effects}
\keyword{internal}