\name{STAS}
\alias{STAS}
\docType{data}
\title{State-Trait Anger Scale (STAS)}
\description{
This dataset contains the scores of the Trait Anger scale of the STAS. The test participants were \eqn{316} first-year psychology students from a university in the Dutch speaking part of Belgium. Participation was a partial fulfillment of the requirement to participate in research. The sample consists of \eqn{73} males and \eqn{243} females, reflecting the gender proportion among psychology students. The average age was \eqn{18.4} years. The data originally come from the package \code{psychotools}, dataset \code{VerbalAgression}. \cr For more info, see \code{https://cran.r-project.org/web/packages/psychotools/psychotools.pdf}.}  
\usage{data(STAS)}
\format{
A \code{data.frame} with \eqn{316} observations on \eqn{3} variables.
  \describe{
    \item{\code{Id}}{The Id number of the student.}
    \item{\code{Gender}}{The gender of the student, coded as a factor.}
    \item{\code{Anger}}{The Trait Anger scale score of the STAS.}
}}
\keyword{STAS}
\keyword{Dataset}