\name{VLT}
\alias{VLT}
\docType{data}
\title{Verbal Learning Test data}
\description{
This dataset contains the Total Recall scores of the Verbal Learning Test (VLT). A total of \eqn{1460} test-participants participated in the study. These are simulated data based on the results described in Van der Elst \emph{et al.} (2005). }  
\usage{data(VLT)}
\format{
A \code{data.frame} with \eqn{1460} observations on \eqn{5} variables.
  \describe{
    \item{\code{Id}}{The Id number of the test participant.}
    \item{\code{Age}}{The age of the test participant (in years).}
    \item{\code{Gender}}{The gender of the test participant, coded as a factor.}
    \item{\code{LE}}{The level of education of the test participant.}
    \item{\code{Total.Recall}}{The Total Recall score. A higher score is indicative for better verbal memory ability.}
}}
\references{
Van der Elst \emph{et al.} (2005). Rey's Verbal Learning Test: Normative data for 1,855 healthy participants aged 24-81 years and the influence of age, sex, education, and mode of presentation. \emph{Journal of the International Neuropsychological Society, 11,} 290-302.

Van der Elst, W. (2023). \emph{Regression-based normative data for psychological assessment: A hands-on approach using R.} New York: Springer. 
}
\keyword{VLT}
\keyword{Dataset}